update standard.tpl to fix too many decimals
this problem causes paypal to capture the money, but the store does not convert the order because of the 3rd decimal place on return from paypal.  This solution seems to fix it once and for all.  The problem occurs when tax is on and you get a thousandth of a penny mathematically (.001)